fastapi-app/requirements.txt
fastapi
uvicorn[standard]
stomp.py
json-log-formatter
